Added js to prevent accidental deletion of post users and catgories

+++ admin/manage-users.php
@@ -113,7 +113,7 @@
                                 <td><?= "{$user['firstname']} {$user['lastname']}" ?></td>
                                 <td><?= "{$user['username']}"?></td>
                                 <td><a href="<?= ROOT_URL ?>admin/edit-user.php?id=<?=$user['id']?>" class="btn sm">Edit</a></td>
-                                <td><a href="<?= ROOT_URL ?>admin/delete-user.php?id=<?=$user['id']?>" class="btn sm danger">Delete</a></td>
+                                <td><a onclick="return confirm('Are you sure that you want to delete <?=$user['username']?>\'s Account?');" href="<?= ROOT_URL ?>admin/delete-user.php?id=<?=$user['id']?>" class="btn sm danger">Delete</a></td>
                                 <td>
                                     <?php if($user['is_admin']== 1):?>
                                         Admin
